+++ CHANGELOG.md
@@ -2,12 +2,12 @@ Changelog
 =========
 
 ### TODO
+- Tiling and offset attaching baked textures?
 - Back button in the import window.
-- Normal strength settings for custom shaders (and baking?).
-- Normal map textures should default to Bump mode when empty.
-- In fact check all the default texture modes...
 
 ### v 0.1.3
+- Normal strength settings for all relevant custom shaders.
+- All normal map, micro normal map and blend normal map textures default to Bump mode when empty.
 - Floating point number conversion in Json file uses correct culture.
 - Logs errors when Json data fails to read.